added ft_memset

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: tylerlover911 <tylerlover911@student.42    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/14 14:28:17 by tylerlover9       #+#    #+#             */
-/*   Updated: 2024/10/17 20:00:05 by mkettab          ###   ########.fr       */
+/*   Updated: 2024/10/18 19:55:20 by mkettab          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include <stdlib.h>
+
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (*s)
